restructure some tests
