+++ integration-test/tests/email-autofill.ios.spec.js
@@ -1,13 +1,12 @@
 import {
     createAutofillScript,
-    defaultIOSScript,
     forwardConsoleMessages,
     setupServer,
     withIOSContext
 } from '../helpers/harness.js'
 import { test as base } from '@playwright/test'
 import {constants} from '../helpers/mocks.js'
-import {emailAutofillPage, loginPage, signupPage} from '../helpers/pages.js'
+import {emailAutofillPage, loginPage} from '../helpers/pages.js'
 import {createWebkitMocks, iosContentScopeReplacements} from '../helpers/mocks.webkit.js'
 
 /**
@@ -38,7 +37,7 @@ test.describe('ios', () => {
         // Load the autofill.js script with replacements
         // on iOS it's the user-agent that's used as the platform check
         await createAutofillScript()
-            .replaceAll(iosContentScopeReplacements)
+            .replaceAll(iosContentScopeReplacements())
             .platform('ios')
             .applyTo(page)
 
@@ -55,44 +54,6 @@ test.describe('ios', () => {
 
         // Because of the mock above, assume an email was selected and ensure it's auto-filled
         await emailPage.assertEmailValue(privateAddress0)
-    })
-    test.describe('Prompting to save from a login form', () => {
-        /**
-         * @param {import("playwright").Page} page
-         */
-        async function setup (page) {
-            await forwardConsoleMessages(page)
-            await createWebkitMocks().applyTo(page)
-            await defaultIOSScript(page)
-            const login = loginPage(page, server)
-            await login.navigate()
-            return login
-        }
-        test('username+password (should prompt)', async ({page}) => {
-            const login = await setup(page)
-
-            const credentials = {
-                username: 'dax@wearejh.com',
-                password: '123456'
-            }
-            await login.submitLoginForm(credentials)
-            await login.assertWasPromptedToSave(credentials)
-        })
-        test('password only (should prompt)', async ({page}) => {
-            const login = await setup(page)
-
-            const credentials = { password: '123456' }
-            await login.submitPasswordOnlyForm(credentials)
-            await login.assertWasPromptedToSave(credentials)
-        })
-
-        test('username only (should NOT prompt)', async ({page}) => {
-            const login = await setup(page)
-
-            const credentials = { username: '123456' }
-            await login.submitUsernameOnlyForm(credentials.username)
-            await login.assertWasNotPromptedToSave()
-        })
     })
     test('autofill a login form', async ({page}) => {
         // enable in-terminal exceptions
@@ -114,7 +75,7 @@ test.describe('ios', () => {
 
         // Load the autofill.js script with replacements
         await createAutofillScript()
-            .replaceAll(iosContentScopeReplacements)
+            .replaceAll(iosContentScopeReplacements())
             .platform('ios')
             .applyTo(page)
 
@@ -122,29 +83,5 @@ test.describe('ios', () => {
         await login.navigate()
         await login.clickIntoUsernameInput()
         await login.assertFirstCredential(personalAddress, password)
-    })
-    test('Prompting to save from a signup form', async ({page}) => {
-        // enable in-terminal exceptions
-        await forwardConsoleMessages(page)
-
-        const {personalAddress} = constants.fields.email
-
-        const credentials = {
-            username: personalAddress,
-            password: '123456'
-        }
-
-        await createWebkitMocks().applyTo(page)
-
-        // Load the autofill.js script with replacements
-        await createAutofillScript()
-            .replaceAll(iosContentScopeReplacements)
-            .platform('ios')
-            .applyTo(page)
-
-        const signup = signupPage(page, server)
-        await signup.navigate()
-        await signup.enterCredentials(credentials)
-        await signup.assertWasPromptedToSave(credentials)
     })
 })
