+++ integration-test/tests/save-prompts.ios.spec.js
@@ -1,0 +1,123 @@
+import {
+    createAutofillScript,
+    defaultIOSScript,
+    forwardConsoleMessages,
+    setupServer,
+    withIOSContext, withIOSFeatureToggles
+} from '../helpers/harness.js'
+import {test as base} from '@playwright/test'
+import {loginPage, signupPage} from '../helpers/pages.js'
+import {createWebkitMocks, iosContentScopeReplacements} from '../helpers/mocks.webkit.js'
+import {constants} from '../helpers/mocks.js'
+
+/**
+ *  Tests for email autofill on ios device
+ */
+const test = withIOSContext(base)
+
+test.describe('Save prompts', () => {
+    let server
+    test.beforeAll(async () => {
+        server = setupServer()
+    })
+    test.afterAll(async () => {
+        server.close()
+    })
+    test.describe('When saving credentials is disabled ❌', () => {
+        test('should not prompt to save', async ({page}) => {
+            await forwardConsoleMessages(page)
+            await createWebkitMocks().applyTo(page)
+
+            /** @type {Partial<FeatureTogglesSettings>} */
+            const toggles = {
+                credentials_saving: false
+            }
+
+            await withIOSFeatureToggles(page, toggles)
+
+            const login = loginPage(page, server)
+            await login.navigate()
+
+            const credentials = {
+                username: 'dax@wearejh.com',
+                password: '123456'
+            }
+            await login.submitLoginForm(credentials)
+            await login.assertWasNotPromptedToSave()
+
+            /**
+             * NOTE: This is here as a sanity check because this test is a negative check
+             * and there are lots of ways this test could pass, but be broken.
+             *
+             * For example if our script fails to load, then this test would normally pass,
+             * but by having this mock check here, it confirms that scripts ran, messages
+             * were sent etc.
+             */
+            await login.assertMockCallOccurred('getAvailableInputTypes')
+        })
+    })
+    test.describe('When saving credentials is enabled ✅', () => {
+        test('Prompting to save from a signup form', async ({page}) => {
+            // enable in-terminal exceptions
+            await forwardConsoleMessages(page)
+
+            const {personalAddress} = constants.fields.email
+
+            const credentials = {
+                username: personalAddress,
+                password: '123456'
+            }
+
+            await createWebkitMocks().applyTo(page)
+
+            // Load the autofill.js script with replacements
+            await createAutofillScript()
+                .replaceAll(iosContentScopeReplacements())
+                .platform('ios')
+                .applyTo(page)
+
+            const signup = signupPage(page, server)
+            await signup.navigate()
+            await signup.enterCredentials(credentials)
+            await signup.assertWasPromptedToSave(credentials)
+        })
+        test.describe('Prompting to save from a login form', () => {
+            /**
+             * @param {import("playwright").Page} page
+             */
+            async function setup (page) {
+                await forwardConsoleMessages(page)
+                await createWebkitMocks().applyTo(page)
+                await defaultIOSScript(page)
+                const login = loginPage(page, server)
+                await login.navigate()
+                return login
+            }
+            test('username+password (should prompt)', async ({page}) => {
+                const login = await setup(page)
+
+                const credentials = {
+                    username: 'dax@wearejh.com',
+                    password: '123456'
+                }
+                await login.submitLoginForm(credentials)
+                await login.assertWasPromptedToSave(credentials)
+            })
+            test('password only (should prompt)', async ({page}) => {
+                const login = await setup(page)
+
+                const credentials = { password: '123456' }
+                await login.submitPasswordOnlyForm(credentials)
+                await login.assertWasPromptedToSave(credentials)
+            })
+
+            test('username only (should NOT prompt)', async ({page}) => {
+                const login = await setup(page)
+
+                const credentials = { username: '123456' }
+                await login.submitUsernameOnlyForm(credentials.username)
+                await login.assertWasNotPromptedToSave()
+            })
+        })
+    })
+})
